requirements.txt
python-dotenv==0.10.1
Django==2.2.7
djangorestframework==3.10.3
ipython==7.9.0
django-jet==1.0.8
django-filter==2.2.0
django-health-check==3.11.1
django-rest-swagger==2.2.0
psycopg2-binary==2.8.4